+++ packages/google_sign_in/google_sign_in_web/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.12.2
 
 * Adds server auth code retrieval to google_sign_in_web.
+* Adds `web_only` library to access web-only methods more easily.
 
 ## 0.12.1